requirements.txt
pandas>=2.1.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
requests>=2.31.0
